backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
bcrypt
numpy
scikit-learn
joblib
pdfplumber
python-docx
httpx
cachetools
python-multipart
openai
pymongo
